Add experiment for new WIP Insights.
After we launch the first set of insights, we will want a way to enable
us to land WIP insights + iterate incrementally without fearing we
accidentally ship them.

This CL introduces the experiment and puts the "Font display" insight
behind it. Note that this is not a comment on the state of this insight,
but it is not included in the first set of 9 insights we are working on
launching, and therefore we want to put it behind the flag for now.

Fixed: 371982243
Change-Id: Ia5d572a09ac12a483da40d906294799862a86367
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5913655
Reviewed-by: Connor Clark <cjamcl@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Connor Clark <cjamcl@chromium.org>