Make 'value interpreter' div optional in memory inspector
We would like to re-use the linear memory inspector in the network
panel to inspect binary content. In that context the value interpreter
doesn't make a lot of sense.

This CL adds a flag to the linear memory inspector component to hide
the value interpreter. The flag is negative ("hide" instead of "show")
so the absence of the flag is consistent with 'false'.

R=kimanh@chromium.org

Bug: 372158884
Change-Id: I1f7c41b770f39929fb295376fd19a1a24e44a96b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5913093
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>