[network] Add RequestBinaryResponseView wrapper around memory inspector
This CL adds a new sub-class of the LinearMemoryInspectorView that
connects "StreamingContentData" (as produced by network requests)
with the linear memory inspector.

The goal is to show this view when inspecting the binary content of
network requests.

The implementation is straight-foward: Everytime a new content chunk
arrives via CDP, we update the view.

R=kimanh@chromium.org

Bug: 372158884
Change-Id: I55c541122f1b0928f6317da9ba2c6f0787b29a15
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5913098
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>