[network] Show memory inspector in the "Response" tab for binary content
This CL changes the Network > Response tab to show the linear memory
inspector for network requests with a binary payload instead of nothing.

Until recently we would blindly decode binary content in some cases
(but not all). This only made sense if the binary content actually
had some valid ascii sequences in it's content as the rest would be
garbage.

This CL restores the ability to inspect binary content, but via the
linear memory inspector which produces a hexdump of the content.
This allows developers to still check ascii sequences in binary
content and even check the binary data directly.

R=kimanh@chromium.org

Fixed: 372158884
Change-Id: Ifdac1607b767a1522343e9e9bfc161a0e5d2ffc3
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5917828
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>