[network] Fix color-scheme in HTML preview for text-only content
This CL fixes a regression introduced by https://crrev.com/c/5785966.
The CL changed how iframes inherit color-scheme: Before iframes would
use the Chrome/OS setting, but now it inherits the color scheme from
the embedding element.

This means if DevTools is in dark mode, then for text-only content
would be "empty" since text color would be white (we always use
a white background regardless of theme for the preview iframe).

In that sense this is not a true regression since this was probably
also a problem before, but is made worse now.

This CL fixes this by always using `color-scheme: light;` for the
preview iframe. This fits with the hardcoded light background color.

R=bmeurer@chromium.org, kprokopenko@chromium.org

Fixed: 371802477
Change-Id: Ia4161c892f1a1695a11b64111723f85d563f559d
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5920234
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>