Do not reappend flamechart after source maps load if not needed
Trace load performance tests started to flake recently due to the
pipeline becoming slower [1]. Upon checking the benchmark graph,
it seemed the recent changes to source mapping support [2] was the
culprit.

After investigating, I realized about the bottleneck: everytime we
attempt to resolve source mappings we reappend the flame chart. We do
this because call hierarchies can change as a consequence of new source
mappings (f.e. if new entries are ignore listed), this can get
expensive with traces with many entries. The problem is not the cost
of the reappending itself, but rather our disposition to always pay it,
without regard for if we need to.

This CL adds a check to the source resolving logic to track whether the
data we display in the flamechart is affected by it. In case it isn't we
don't dispatch a `SourceMappingsUpdated` event to prevent an unnecessary
rerender.

Longer term follow up work could check with more detail if call
hierarchies are affected, so that we only redraw the flamechart if
possible (instead of reappending + drawing). Also, we could change the
appending system so that a re-appending can be done for individual tracks
and not only the whole timeline.

[1] https://devtools-frontend-perf.corp.goog/e/?begin=1725366900&end=1728477245&queries=stats%3Dmean%26test%3DTraceLoad&requestType=0&selected=commit%3D15112%26name%3D%252Cbot%253DStand-alone%2520Linux%252Cmaster%253Dclient.devtools-frontend.integration%252Cstats%253Dmean%252Ctest%253DTraceLoad%252Cunits%253Dms%252C

[2] https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5882655

Fixed: 370879877
Change-Id: Ic1355d098d4d78f5461b5208b257cfb663db2379
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5920399
Commit-Queue: Andres Olivares <andoli@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>