DevTools: Add ExtensionStorageItemsView and show in Application panel
Adds an ExtensionStorageItemsView and shows this when selecting an
extension storage area in the Application panel. This provides the UI
for editing values in extension storage.

Also added logic to ExtensionStorageModel to only add a storage area
once we have checked it is accessible. We agreed this was preferable to
adding a new CDP method purely for this purpose.

Currently, it is only possible to edit values and set a string value. A
follow-up CL will add support for JSON based objects.

Bug: 40963428
Change-Id: I04da2f49afcf40081e0c5e5a0ca5c82db6b68406
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5918627
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Oliver Dunk <oliverdunk@chromium.org>