[deps] Upgrade TypeScript to 5.6.2
This upgrade requries some changes to the code base. TS 5.6 made the
types for the iterator protocol more strict. We have some places in the
code base where we retrieve the first item of `Set` via the `values()`
iterator. This CL adds 'as' casts or uses 'done' as appropriate. For
test code we use '!'.

R=alexrudenko@chromium.org

Bug: None
Change-Id: I09f1dccb31485a7e3036870f7688aa57f1202018
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5920100
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>