RPP: fix network throttling w/ custom presets
This CL fixes issues with the network throttling in the Performance
panel and how it interacts with custom presets.

The code relies on the presence of an i18nKey to uniquely identify
presets, this works for the ones we define but not for custom user ones.
This CL updates the code to generate a key by either usign the i18nKey
or falling back to the title.

This fixes the issues seen in the attached bug where selecting one
custom preset selects them all, and the value is not synced correctly
across to the network panel.

Additionally this CL adds a `title` attribute to each menu item, else
the tooltip for each item is `Network throttling: X` which shows the one
you happen to have selected.

Fixed: 370332091
Change-Id: I7ce068df9043b69e3ac890a0f24197272415d2c0
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5923562
Commit-Queue: Adam Raine <asraine@chromium.org>
Reviewed-by: Adam Raine <asraine@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>