DevTools: Don't use tree view in extension storage viewer for top frames
When viewing a top-level chrome-extension:// page, it makes sense to
have the extension storage viewer sidebar entries use the same layout
as an extension service worker.

Achieve this by simplifying the logic to determine if the tree view is
used to just look at if the storage is for the same origin as the top
frame.

Bug: 40963428
Change-Id: I5c17d642cf2d8e0e06fca7278e987c3c8bbaa1d2
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5920619
Commit-Queue: Oliver Dunk <oliverdunk@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>