DevTools: Use extension ID as fallback in extension storage tree
In the ideal case, the Extension Storage section in the Application
panel should use the extension name as the title of the parent element.
However, this comes from the execution context name which is not always
set.

As cases are identified, these are being fixed, but add a fallback to
the extension ID to make sure we never show an empty string as the
title.

Bug: N/A
Change-Id: I5e77758fcd28e1b06d43e23b40017f3627ead811
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5920102
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Oliver Dunk <oliverdunk@chromium.org>