Insights: fix invalid LCP phases
Although we do not expect it for new traces, older trace files can be
missing some data that we need to calculate the phases, the result of
which is `NaN` appearing as a calculated phase value.

This CL ensures that if we encounter this, we bail and return
`undefined` for the phases.

Fixed: 352413429
Change-Id: I16e2239ad80a8d63250592542c514465fad610d0
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5924437
Reviewed-by: Connor Clark <cjamcl@chromium.org>
Commit-Queue: Connor Clark <cjamcl@chromium.org>