Use polling to update FS, CI, and AI innovation settings on user login
Console insights and Freestyler cannot be used, and AI settings are
disabled if the user is not logged in. Previously the user had to
restart DevTools after logging in for everything to be updated. This
CL adds a polling mechanism to the relevant components, so that after
logging in the DevTools state is updated automatically and the
features can be used without restarting DevTools.

Bug: 370004134, 371530995, 372632942, 370808543
Change-Id: If99c61027887b85b1e03cbda65ab3f8152b1123c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5923556
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Wolfgang Beyer <wolfi@chromium.org>