[RPP] Make 'EntryBorderOutlineType' only the visible width of the entry
Set the width of the Entry Border Outline that is used for multiple overlays to be the width of the visible part of the entry instead of the whole entry.

This solves multiple issues:

1. Sometimes it was not possible to add a label annotation because only part of the entry was displayed on the flame chart, and the label, which is centered in the middle of the entry, was not visible (https://screenshot.googleplex.com/FZhEt5jTUHvS9SR).
Now the label stays visible for as long as possible because it will be centered on the visible part of the entry.

This change resolves the annotation creation issue and also ensures the label remains visible during navigation for as long as possible.


2. Similarly, the arrow for Connection annotation will point to a visible part of the entry for as long as possible.


3. The entry wrapper on all overlays does not longer overlap the scrollbar

Before video: http://screencast/cast/NTAzODcwNDY4MTgxMTk2OHwwYjYyMTM2Ny05OA

After video: http://screencast/cast/NjIxNDgxNDU2ODI4NDE2MHxiZWRhMzlkMy03NQ


Bug: 371044304,372899466
Change-Id: I6b1ec89e916e4ffd41216e1dc1d63462c2803120
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5921740
Auto-Submit: Alina Varkki <alinavarkki@chromium.org>
Commit-Queue: Alina Varkki <alinavarkki@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>