RPP: fix revealing outside of breadcrumbs
This CL fixes an issue where revealing an annotation / insight outside
of the breadcrumbs.

The TraceBounds service will bail if the new window it is given matches
the existing one. The problem was I had moved this check to after we
adjust the window for minimap bounds. This caused a problem where when
an insight/annotation was revealed that is outside of the current
breacrumb, because the check that previously returned `true` now
returned false, because the code would adjust the new window and then
return `false`. This lead to a slight race condition where the
TraceBounds is updating as a result of the user clicking the
insight/annotation, but at the same time the TimelineOverview component
would dispatch an event that the window had changed too.

Moving the check earlier (I actually keep them both now, we might as
well eagerly look for any chance to not do work) fixes this problem and
ensures the timeline updates correctly.

As a drive-by, I also added tests for windowFitsInsideBounds, which we
thought might be the problem. It turned out not to be, but we should
have added unit tests when we landed the helper, so I added them now.

Bug: none
Change-Id: Ie88b062f0df55e8c3f43e68d8d6dc9fcdfc11108
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5928645
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Alina Varkki <alinavarkki@chromium.org>
Commit-Queue: Alina Varkki <alinavarkki@chromium.org>