[RPP] Improve ignorelist detection
So far the only condition for ignore listing urls in the panel was the
matching against regex patterns added by the user in the ignore list
settings. This is because the devtools bindings utils don't extract
details about a code location that could be used to determine if it
should be ignored. Instead, they rely on the caller passing those
details along with the location being tested.

So, the following ignore list conditions are implemented by this CL by
passing the relevant code location details to the bindings utils:

1. Script is contained in the "ignoreList" field of the source map.
   (aka script is a "known third parties").
2. Script is a content script.

With this in place, our ignore listing behaviour should be consistent
with the rest of devtools.

Fixed: 358572661
Change-Id: Id36ede5873e504b83546bdfeb32c4e0f053ac48b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5928964
Commit-Queue: Andres Olivares <andoli@chromium.org>
Reviewed-by: Alina Varkki <alinavarkki@chromium.org>