[har] Correctly preserve port information.
This aligns our HAR import/export with Firefox, and stores the remote
port into the "connection" field of the HAR entry, and the Chrome
specific connectionId into a custom "_connectionId" field. The spec[^1]
doesn't strictly require this, but hints that the "connection" field
is a good place to store the (server) port, and since Firefox already
does it this way, we can just as well follow suit.

[^1]: http://www.softwareishard.com/blog/har-12-spec/

Bug: 369859833
Change-Id: Ib2cfe8e30da745ce91e88fbc5768b68cef2e5403
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5938877
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>