[screencast] Fix broken screencast toggle.
This broke as part of https://crrev.com/c/5563488, which switches the
toolbar buttons to use the button component. As part of this the logic
to toggle upon clicks is already built into the toolbar button now,
and the additional code to flip the toggle in the "click" handler
caused the toggle to immediately go back to the previous state.

Fixed: 373263964
Bug: 325441197
Change-Id: I9a2bc5e801e1c1e69f584226574ddb1632c0dec5
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5939707
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Yang Guo <yangguo@chromium.org>
Reviewed-by: Yang Guo <yangguo@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>