[ui] Make TreeOutline listen to "keyup" instead of "keydown".
The primary motivation for this is to allow the processing of "click"
events generated from "Enter" key presses in elements within a
TreeOutline. Before this CL, the "keydown" handler in the TreeOutline
would always consume the "Enter" key before the child element gets a
chance to turn that into a "click" event.

There's a secondary reason however, which is that up-events should be
preferred for accessibility reasons, see [^1] which describes this in
the context of so-called pointer device buttons, but the same applies
to keyboards as well.

[^1]: https://www.w3.org/WAI/WCAG21/Understanding/pointer-cancellation

Fixed: 373168872
Change-Id: I3b82e1fde221383279ee6a690af3e436fcc98033
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5939709
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Changhao Han <changhaohan@chromium.org>
Commit-Queue: Changhao Han <changhaohan@chromium.org>