RPP: add test for details view showing a frame
I want to start refactoring some of the `TimelineSelection` logic, but
realised we are missing tests for a few places where it is used. This CL
adds a test to ensure screenshots are shown when the user selects a
frame, and moves the rendering logic into a standalone function to make
it a bit easier to read.

Bug: 370716587
Change-Id: Iedb0de633884ddc2b0b2c0a74c9a98c6de315870
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5946889
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>