[sources] Check media type for images and fonts
If users navigate a chrome tab to an image/font it has ResourceType
'Document' and currently we'll show an error (or raw bytes).

Instead we can be a bit more permissive and attempt to render an
image/font if the media type indicates as such.

R=bmeurer@chromium.org

Fixed: 373728013
Change-Id: Ia931cbacfe0b0f41ba1789214c683e9857dd1b5b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5952209
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>