Make invisible issues
The new 3p related cookie issues do not need to be surfaced to the
user in the issue pane. Instead, they are going to be shown in the
new Cookie Report. They still need a description to be properly
added by the issuesManager, so this is why a placeholder description
was added. The issueAggregator will exclude these issues so that
the issue pane does not show them, while still making the issues
accessible via the issuesManager.

Bug: 365737493
Change-Id: Ib7c4b918935b7604cf57b0a1dcebb90c0297f4b2
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5950123
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Joshua Thomas <masnoble@chromium.org>
Reviewed-by: Shuran Huang <shuuran@chromium.org>