[Freestyler] Create History from run yields
We want to update the way we handle history so we can restore it from storage eventually.
For this we store the whole run with every detail, that way we can rebuild both the UI and the AIDA history from a single point.

go/chrome-devtools:ai-history

Bug: 369303513
Change-Id: Ic97e3222d4f7827b900d904e73d607e72391265d
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5954344
Commit-Queue: Nikolay Vitkov <nvitkov@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>