[RPP] Desaturate unrelated events when insight/search is active
This is added initially as an experiment, as there is a long tail of
improvements to make this UX treatment seamless.

Adds to FlameChart the ability to transform the colors used for drawing
based on a set of currently relevant events. The active insight and the
searching feature both utilize this now.

FlameChart now maintains a "color => grayscale color" mapping.
"getColorForEntry" is used to select a color that takes into account the
currently configured relevant events.

Appenders have their own custom drawing, and so are handled by passing
them a "tranformColor" callback, which plugs into FlameChart's logic for
determining if an event needed to be dimmed.

The active insight derives which events are relevant from the provided
overlays. Events associated with an overlay are relevant, as well as
events that fall within the time range of an overlay that defines a
range.

Drive-by change: the LCPPhases insight now produces an overlay for the
LCP request event, if any.

https://i.imgur.com/BS8ANMc.png

Bug: 370836271
Change-Id: I832405a439b9b631485be2e17465e1bfadbe56e1
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5939713
Reviewed-by: Paul Irish <paulirish@chromium.org>
Commit-Queue: Connor Clark <cjamcl@chromium.org>