Categorize heap snapshot objects by constructor name and location
Currently, JS objects in the heap snapshot Summary view are categorized
according to the names of their constructors. However, it is possible
for distinct constructors to have the same name, meaning unrelated
objects end up in the same group. With this change, JS objects are
categorized together according to a combination of their constructor
names and source locations. This combination of name and location is
referred to as a "class key" throughout the code.

Bug: 357902505
Change-Id: Idcd9a4904aa739794eabbb14190b6c63156098f0
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5943606
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Seth Brenith <seth.brenith@microsoft.com>