Fix scrolling into view in Treeoutline
The scrollable container for the Treeoutline has changed, and as a
result, the wrong container was used for scrolling when we attempted
to scoll an item in the tree into view.

This CL changes the way we retrieve the correct scroll container.
Instead of using the parent element, we either get the parent
element or the shadow host.

Fixed: 374592181
Change-Id: Ib3d033b0a4d1c7f7b88e2af63f9bd1f67de452d9
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5961819
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>