RPP: Introduce AICallTree, rewrite of TreeHelper AINode
Revamp of the data structure and serialization format of a flame chart call tree, being passed to Freestyler. This impl leverages TimelineProfileTree for faster/easier tree construction and manipulation. We build our tree with only the events from the same track that overlap the selected event.

Serialization: We switch from JSON to a YAML-like format that's 35% more efficient with the tokenizer. A few variants were tested and this format was chosen to be token-efficient and meaningfully clear to the LLM. Notably:
 - rather than nested children, we use adjacency lists.
 - URLs are particularly token-inefficient. To avoid repeating, we output a lookup table.

Within the Freestyler chat UI, the Call frame chip is now clickable. It relies on a new Revealable class: SDK.TraceObject.RevealableEvent

Bug: 370436840,372658699
Change-Id: I89baabfd9e7b1fdd183a6a9ea12bf2a318125fbc
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5963059
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Auto-Submit: Paul Irish <paulirish@chromium.org>