[Animations] Auto start animation model
For adding an animation link swatch from styles tab to the animations
panel, we need to have a source of truth regarding captured animations
between styles tab and the animations panel so that we show the
links only for the animations that are captured.

Automatically starting AnimationModel enables us to capture animations
& store them even before the animations panel is opened.
That way, in the styles tab, we know the captured animations and
can link to them in the Animations tab.

Bug: 349566091
Change-Id: I1c2d7bbc811a4d6bef2b0a7a9bb1738da23db30c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5966113
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>