Do not auto focus within Ignore Listing settings
With the new UI design, the button to add a new pattern is at the
end of the settings page. We don't want to automatically scroll
to the end of the settings page, so removing the auto focus here.

Bug: 368238837
Change-Id: I164c74b3314321169418d06f46bfc6d09ce1d734
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5971497
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>
Reviewed-by: Eric Leese <leese@chromium.org>
Auto-Submit: Kim-Anh Tran <kimanh@chromium.org>