[network] Move RequestBinaryResponseView to different module
This CL renames the "RequestBinaryResponseView" to
"StreamingContentHexView" and moves it to "source_frame" components.

We want to re-use the linear memory inspector in the
BinaryResourceView(Factory) and importing the network panel there would
be a layering violation.

At the same time, we also can't depend on the linear memory inspector
panel directly as well, as that would also be a layering violation.
To solve this, this CL re-implements the "LinearMemoryInspectorView"
in a way that doesn't have a dependency on the
"LinearMemoryInspectorController".

In a follow-up CL the "StreamingContentHexView", will replace the
"hex" view portion of the "BinaryResourceView(Factory)". The utf-8 and
base64 will stay as-is modulo supporting streaming content.

R=kimanh@chromium.org

Bug: 375546679
Change-Id: If2f3ce62e8f72bd29d207b998aba17613d3b8237
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5972706
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>