[network] Always show "Open in Network panel" context menu item.
Previously we'd only show the "Reveal in Network panel" context menu
item when we had an actual network request recorded for the resource.
While this is technically correct, it leaves the developer wondering
why the menu item isn't showing (the reasons are subtle). So with this
CL we rename the item to "Open in Network panel", and always show it,
but when there's no resource, we show it disabled and with a note
"(missing request)" appended.

Fixed: 372873222
Screenshot: https://i.imgur.com/elaOkCC.png
Change-Id: I8bf64827145aca93bbd2a0def6c22b9e3149315f
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5975683
Reviewed-by: Peter Müller <petermueller@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>