[Freestyler] Update the layout and make `main` own the scroller
After moving the `.input-form` inside `main` container,
we switch between it overlaying the messages (with
`position: sticky`) or rendering it after the messages
depending on the `main`s height.

The layout now is:
* `main` now owns the scroller which makes the scroller span
to the end of the `.input-form`.
* when there is empty space between left and right edges of the messages
container and input form (main width > 688px), appearing of scrollbar
is not causing the layout to shift.
* when there isn't empty space between left and right edges of the
messages container and input form (main width <= 688px), appearing
of scrollbar causes the content to jump from right.
* when there isn't empty space between left and right edges of the
messages container, the lines of the message touch the edges.
* the edges of the message & input form is always aligned. (screenshot crbug.com/40248266#comment38)

Drive-by:
* There is now less top padding in `.input-form`.

Fixed: 372317501, 373600025
Change-Id: Id65d56bf7a424c6c3794180e475902221b8ade4e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5975457
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>
Reviewed-by: Wolfgang Beyer <wolfi@chromium.org>