RPP: rewrite TimelineSelection class
The `TimelineSelection` class has become confusing to work with since we
completed the data migration to the new trace engine.

It differentiated selections of types that used to be different concepts
in the codebase: frames, trace events and network events. Now, these are
all defined as `Trace.Types.Events.Event`, so we don't need each to have
its own "selection type".

We have all the methods in `Trace.Types.Events` that let you determine
the type of event, so rather than duplicate some of those in
`TimelineSelection`, we now categorise a selection as one of two things:

1. an event
2. a time range

Once you determine your selection is an event, you can use all the same
helpers in `Trace.Types.Events` to determine the exact type of event, if
you need to.

Fixed: 370716587
Change-Id: I256a6841af280a1eff4cecacc39a9f351b3ebc11
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5975681
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>