Fix Python SyntaxWarning in //scripts/build/wasm-as.py.
In Python 3.12, unrecognized escape character sequences in a string
produce a "SyntaxWarning" log message when the file is parsed. The
fix is (usually) to change the string from a regular string to a raw
string, by putting an "r" in front of the leading quote.

Bug: 376746798
Change-Id: Ie888a8de6ae42c4daae6fa5028134722634d70c1
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5981155
Commit-Queue: Dirk Pranke <dpranke@google.com>
Reviewed-by: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Liviu Rau <liviurau@chromium.org>