[button] Ensure that Enter/Space triggers click events.
Previously we have applied specific duck tape in a bunch
of places to prevent "keydown" event handlers on parent
elements from eating up Enter/Space events and thus not
firing click events on buttons as expected. With this CL
we put the logic solely into the `<devtools-button>`
component (together with a test) to ensure not every use
site has to deal with this, and it's less easy to miss.

Fixed: 373168872, 372411090
Change-Id: I00c183547a7f7f2e20a95467687885db463280ca
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5982917
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>