Simplify copy-to-clipboard helpers in BinaryResourceViewFactory
This CL replaces the utf8/hex methods on BinaryResourceViewFactory
with a slower but sync version and removes the caching. This makes
handling with the "StreamingContentData" easier since we don't have
to update the cached value when the "StreamingContentData" updates.

Note that this is not a big deal performance wise. The utf8/hex
helpers are not invoked often: When the user triggers a
"copy-to-clipboard" action and for utf8 when we create the view
initially. Using the sync version over `fetch` is also slightly
slower (by a factor of 2x for a 20MB string), but the sync code
is much simpler to deal with.

R=pfaffe@chromium.org

Bug: 375546679
Change-Id: Ie5b6038cb1d0183014058c5b0bdef76c6bec0ca4
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5993270
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>