Handle StreamingContentData updates in BinaryResourceViewFactory
This is the last piece of the puzzle: Update the ResourceSourceFrame
views that the BinaryResourceViewFactory produces when the content
updates.

The "cleaner" approach would be to pass a custom "ContentProvider"
to "ResourceSourceFrame" that spits out a wrapped
"StreamingContentData" that applies the utf8/base64 transformation to
the original "StreamingContentData". That is because the
"ResourceSourceFrame" actually knows how to subscribe to
"StreamingContentData" updates. This is a lot of boilerplate code
though and not very readable, so the next best thing is this CL.

R=pfaffe@chromium.org

Fixed: 375546679
Change-Id: I982859fd255354ea8a73cc928bc34640ec78a87b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5993272
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>