[Freestyler] Max timeout of Observation
Race the Observation with a timeout Promise and report the correct
error to the LLM so that it can try to recover.
This mitigates the issue of SideEffect's and Observations getting
stuck due to Promises that never resolve.

Fixed: 373893673
Change-Id: I209c3b07244bd8412fce5e4c0e47e4af1fbf6f77
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5993651
Commit-Queue: Nikolay Vitkov <nvitkov@chromium.org>
Auto-Submit: Nikolay Vitkov <nvitkov@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>