Clean up the code for `OverviewGrid.ts`
This file has three kinds of value to describe the start/end of a window, one is the start pixel, one is the start ratio and one is the start time/byte. Which is pretty confusing, so this CL rename the variable to make the code more understandable.

Also the two classes here will be used in Performance panel and the Heap allocation in Memory panel, so add some comments to make it clearer.

Bug: none
Change-Id: I6ff11f1469d5ce715326763f895df425181d01fb
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5994170
Commit-Queue: Nancy Li <nancyly@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>