[AI Assistance] Auto select agents for open panels
This CL updates the behavior to auto select an agent
based on the open panels.

In addition to that, updates `new chat` and `delete
chat` buttons to auto select agents too. Previously
it was selecting the last agent and now it'll select
the agent based on the open panel.

Fixed: 376431804
Change-Id: I2a5dda87c7364ac4d1f6a7521e45a9a6d672d8f4
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5999478
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>