DevTools: Enable extension storage in Application panel by default.
Enables the Application > Extension Storage feature by default by
removing the experiment. The experiment is currently available in the
early stable release and the feedback has been largely positive.

There is still follow up work (for example, we didn't add the CDP
methods which would allow us to automatically refresh the view when data
changes). However, this shouldn't need to block the first release and
there is a refresh button which is consistent with other parts of
DevTools.

Bug: 40963428
Change-Id: I9c8d949b1499d0a7b23a8a20174c88c702cdacf0
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6005076
Commit-Queue: Oliver Dunk <oliverdunk@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>