[main] Don't switch to default ExecutionContext for workers on creation
This CL prevents the "ExecutionContextSelector" to switch the "Target"
flavor when worker contexts are created.

The "Sources" panel uses "Target" UI.Context flavor to determine which
target to show. When a target is paused, the "Sources" panel switches
the "Target" UI.Context flavor over to the paused target. This is
what users expect.

The problem is that Chrome allows individual targets to be paused,
while others keep on running. This means while one target may hit
a breakpoint, others are created and start running.

Now, if we get a "ExecutionContextCreated" event AFTER we paused,
we might switch targets again, and switch away from the paused
target. This is unintuitive for users, because it looks like
breakpoints/debugger statements are not hit, because the "Sources"
panel shows the wrong target (even though there is another target
thats paused).

This CL works around the problem by not switching the
"ExecutionContextSelector" when worker ExecutionContexts are
created. These are mostly uninteresting anyway and we should only
switch to worker targets when we either pause in them, or the
user selects them explicitly.

R=dsv@chromium.org

Fixed: 373893057
Change-Id: I5dfc15fb81acbc9d6a073b0f4b540dc557c524fc
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/5999197
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>