Context menu: fix event handling on close
On Windows, closing a context menu by clicking on the button which
opened it, would immediately re-open the menu. A fix was added in
https://crrev.com/c/5592569, which relies on `requestAnimationFrame`.
rAF does not trigger when when the window is not visible, which can
happen for the device toolbar (rendered in the main browser window)
when DevTools is detached and hidden behind another window.

This CL replaces `requestAnimationFrame` with `setTimeout`.

Bug: 377161958, 339560549
Change-Id: Id6af436c41d14a6e16160faff4e160e40bbe4122
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6004964
Auto-Submit: Wolfgang Beyer <wolfi@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>